Move swing detectors out of indicator namespace
