+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## Unreleased
 
 ### Added
-- **Unified recent highs/lows with a canonical default**: Added `RecentSwingIndicators` paired factories for fractal, ATR ZigZag, adaptive ZigZag, slope-change, prominence, and consensus methods. The canonical default is an OHLC-aware ATR(14) ZigZag, while a separate status-aware point view exposes forming terminal extremes without changing confirmed-only indicator semantics. New bounded prominence indicators and detector support identify structurally salient peaks/troughs, swing configuration records live with the swing detector API instead of the first-class indicator namespace, and ossified Coinbase ETH/USD fixtures compare every method from five-minute through daily bars.
+- **Unified recent highs/lows with a canonical default**: Added `RecentSwingIndicators` paired factories for fractal, ATR ZigZag, adaptive ZigZag, slope-change, prominence, and consensus methods. The canonical default is an OHLC-aware ATR(14) ZigZag, while a separate status-aware point view exposes forming terminal extremes without changing confirmed-only indicator semantics. New bounded prominence indicators and detector support identify structurally salient peaks/troughs, swing detectors/configuration records live under `org.ta4j.core.analysis.elliott.swing` instead of the first-class indicator namespace, and ossified Coinbase ETH/USD fixtures compare every method from five-minute through daily bars.
 - **Concise strategy JSON authoring with `Strategy.fromJson(...)` (`CF-113`)**: You can now hand ta4j an opt-in `version: 2` strategy envelope instead of writing the full descriptor tree, which makes quick SMA/EMA crossover setups, RSI threshold rules, stop-gain/stop-loss exits, and short-starting strategies much easier to generate or edit. The bounded shorthand accepts `entryRule` / `exitRule`, `AndRule` / `OrRule`, shorthand indicators like `ClosePrice`, `SMA(...)`, `EMA(...)`, and `RSI(...)`, then validates inputs strictly and normalizes everything back to canonical descriptor JSON internally.
 - **Unified named-asset shorthand for v2 serialization**: Added immutable `NamedAssetRegistry` / `NamedAssetProvider` APIs and compact function-call shorthand for indicators, rules, strategies, and analysis criteria. Existing canonical `toJson()` output remains unchanged, while opt-in helpers such as `Strategy#toCompactJson()`, `Strategy#fromExpression(...)`, `Rule#fromExpression(...)`, `Indicator#fromExpression(...)`, and `AnalysisCriterion#fromExpression(...)` support compact authoring and CLI-friendly criterion names like `NetProfit` and `SharpeRatio`.
 - **Lossless analysis criterion descriptors**: `AnalysisCriterion#toJson()` now preserves supported constructor state such as return representation and boolean flags, and rejects unsupported stateful criteria instead of silently serializing them as default instances.
