+++ ta4j-core/src/main/java/org/ta4j/core/analysis/elliott/package-info.java
@@ -1,0 +1,14 @@
+/*
+ * SPDX-License-Identifier: MIT
+ */
+/**
+ * Elliott Wave analysis support APIs.
+ *
+ * <p>
+ * This package is for non-indicator Elliott support types such as detector
+ * interfaces, configuration records, filters, and runner-facing analysis
+ * helpers. First-class per-index Elliott outputs remain regular
+ * {@code *Indicator} implementations under
+ * {@code org.ta4j.core.indicators.elliott}.
+ */
+package org.ta4j.core.analysis.elliott;
