+++ ta4j-core/src/main/java/org/ta4j/core/analysis/elliott/swing/FractalSwingDetector.java
@@ -1,7 +1,7 @@
 /*
  * SPDX-License-Identifier: MIT
  */
-package org.ta4j.core.indicators.elliott.swing;
+package org.ta4j.core.analysis.elliott.swing;
 
 import java.util.List;
 import java.util.Objects;
