+++ ta4j-core/src/main/java/org/ta4j/core/analysis/elliott/swing/SwingPivotType.java
@@ -1,7 +1,7 @@
 /*
  * SPDX-License-Identifier: MIT
  */
-package org.ta4j.core.indicators.elliott.swing;
+package org.ta4j.core.analysis.elliott.swing;
 
 /**
  * Pivot classification for swing points.
