+++ ta4j-core/src/main/java/org/ta4j/core/analysis/package-info.java
@@ -9,7 +9,8 @@
  * {@link org.ta4j.core.analysis.CashFlow CashFlow}, window definitions and
  * context policies for window-aware criterion evaluation
  * ({@link org.ta4j.core.analysis.AnalysisWindow AnalysisWindow},
- * {@link org.ta4j.core.analysis.AnalysisContext AnalysisContext}), and to
- * calculate {@link org.ta4j.core.analysis.cost.CostModel trading costs}.
+ * {@link org.ta4j.core.analysis.AnalysisContext AnalysisContext}), to calculate
+ * {@link org.ta4j.core.analysis.cost.CostModel trading costs}, and to host
+ * non-indicator analysis support APIs such as Elliott swing detectors.
  */
 package org.ta4j.core.analysis;
