+++ ta4j-core/src/main/java/org/ta4j/core/indicators/RecentSwingIndicators.java
@@ -10,13 +10,13 @@ import java.util.Optional;
 import org.ta4j.core.Bar;
 import org.ta4j.core.BarSeries;
 import org.ta4j.core.Indicator;
-import org.ta4j.core.indicators.elliott.swing.AdaptiveZigZagConfig;
-import org.ta4j.core.indicators.elliott.swing.ProminenceSwingConfig;
-import org.ta4j.core.indicators.elliott.swing.SlopeChangeConfig;
-import org.ta4j.core.indicators.elliott.swing.SwingDetector;
-import org.ta4j.core.indicators.elliott.swing.SwingDetectors;
-import org.ta4j.core.indicators.elliott.swing.SwingPivot;
-import org.ta4j.core.indicators.elliott.swing.SwingPivotType;
+import org.ta4j.core.analysis.elliott.swing.AdaptiveZigZagConfig;
+import org.ta4j.core.analysis.elliott.swing.ProminenceSwingConfig;
+import org.ta4j.core.analysis.elliott.swing.SlopeChangeConfig;
+import org.ta4j.core.analysis.elliott.swing.SwingDetector;
+import org.ta4j.core.analysis.elliott.swing.SwingDetectors;
+import org.ta4j.core.analysis.elliott.swing.SwingPivot;
+import org.ta4j.core.analysis.elliott.swing.SwingPivotType;
 import org.ta4j.core.indicators.helpers.HighPriceIndicator;
 import org.ta4j.core.indicators.helpers.LowPriceIndicator;
 import org.ta4j.core.indicators.zigzag.RecentZigZagSwingHighIndicator;
