+++ ta4j-core/src/test/java/org/ta4j/core/analysis/elliott/swing/AdaptiveZigZagSwingDetectorTest.java
@@ -1,7 +1,7 @@
 /*
  * SPDX-License-Identifier: MIT
  */
-package org.ta4j.core.indicators.elliott.swing;
+package org.ta4j.core.analysis.elliott.swing;
 
 import static org.assertj.core.api.Assertions.assertThat;
