+++ ta4j-core/src/test/java/org/ta4j/core/indicators/RecentSwingIndicatorsTest.java
@@ -15,11 +15,11 @@ import org.ta4j.core.indicators.RecentSwingIndicators.Confirmation;
 import org.ta4j.core.indicators.RecentSwingIndicators.Method;
 import org.ta4j.core.indicators.RecentSwingIndicators.Pair;
 import org.ta4j.core.indicators.RecentSwingIndicators.SwingPoint;
-import org.ta4j.core.indicators.elliott.swing.SwingDetector;
-import org.ta4j.core.indicators.elliott.swing.SwingDetectorResult;
-import org.ta4j.core.indicators.elliott.swing.SwingDetectors;
-import org.ta4j.core.indicators.elliott.swing.SwingPivot;
-import org.ta4j.core.indicators.elliott.swing.SwingPivotType;
+import org.ta4j.core.analysis.elliott.swing.SwingDetector;
+import org.ta4j.core.analysis.elliott.swing.SwingDetectorResult;
+import org.ta4j.core.analysis.elliott.swing.SwingDetectors;
+import org.ta4j.core.analysis.elliott.swing.SwingPivot;
+import org.ta4j.core.analysis.elliott.swing.SwingPivotType;
 import org.ta4j.core.indicators.zigzag.RecentZigZagSwingHighIndicator;
 import org.ta4j.core.indicators.zigzag.RecentZigZagSwingLowIndicator;
 import org.ta4j.core.mocks.MockBarSeriesBuilder;
