+++ ta4j-core/src/test/java/org/ta4j/core/indicators/elliott/ElliottWaveAnalysisRunnerTest.java
@@ -22,9 +22,9 @@ import org.ta4j.core.BarSeries;
 import org.ta4j.core.BaseBarSeries;
 import org.ta4j.core.indicators.elliott.confidence.ConfidenceModel;
 import org.ta4j.core.indicators.elliott.confidence.ElliottConfidenceBreakdown;
-import org.ta4j.core.indicators.elliott.swing.SwingDetector;
-import org.ta4j.core.indicators.elliott.swing.SwingDetectorResult;
-import org.ta4j.core.indicators.elliott.swing.SwingFilter;
+import org.ta4j.core.analysis.elliott.swing.SwingDetector;
+import org.ta4j.core.analysis.elliott.swing.SwingDetectorResult;
+import org.ta4j.core.analysis.elliott.swing.SwingFilter;
 import org.ta4j.core.mocks.MockBarBuilderFactory;
 import org.ta4j.core.mocks.MockBarSeriesBuilder;
 import org.ta4j.core.num.NumFactory;
